+++ CHANGELOG.md
@@ -1,2 +1,11 @@
+## 0.1.3
+fix share menu position
+## 0.1.2
+fix homepage
+## 0.1.1
+add description
+## 0.1.0
+add macos compatibility
 ## 0.0.2
 add message type cast to kotlin
+